refactor: group relic overlay components
