+++ src/App.tsx
@@ -73,13 +73,13 @@ import MarketHelper from "./MarketHelper";
 import RelicHelper from "./RelicHelper";
 import RivenAnalyzer from "./riven/RivenAnalyzer";
 import RivenOverlayWindow from "./riven/RivenOverlayWindow";
-import RelicPickOverlay from "./RelicPickOverlay";
+import RelicPickOverlay from "./relic-overlay/RelicPickOverlay";
 import TimerHelper, { fmtMs } from "./TimerHelper";
 import { useWorldState } from "./worldstate";
 import { notify } from "./lib/notify";
 import { collectNewMatches } from "./fissureAlerts";
 import Statistics from "./Statistics";
-import Overlay from "./Overlay";
+import Overlay from "./relic-overlay/Overlay";
 import ModularWindow from "./modular-window/ModularWindow";
 import ModularWindowPage from "./modular-window/ModularWindowPage";
 import SettingsModal from "./SettingsModal";
