+++ src/relic-overlay/Overlay.tsx
@@ -2,13 +2,13 @@ import { useState, useEffect, useRef } from "react";
 import { listen } from "@tauri-apps/api/event";
 import { invoke } from "@tauri-apps/api/core";
 
-import { overlayScale } from "./lib/uiScale";
-import { PREFERENCE_KEYS } from "./constants/preferences";
-import { TAURI_COMMANDS, TAURI_EVENTS } from "./constants/tauri";
-import type { CraftingJob, QuantityMap, ShallowRecipeComponent } from "./types/items";
-import type { RelicOverlayPriority } from "./types/settings";
-import type { PendingRelicRewards, RelicRewardsPayload } from "./types/tauri";
-import type { InventoryUpdate } from "./types/inventory";
+import { overlayScale } from "../lib/uiScale";
+import { PREFERENCE_KEYS } from "../constants/preferences";
+import { TAURI_COMMANDS, TAURI_EVENTS } from "../constants/tauri";
+import type { CraftingJob, QuantityMap, ShallowRecipeComponent } from "../types/items";
+import type { RelicOverlayPriority } from "../types/settings";
+import type { PendingRelicRewards, RelicRewardsPayload } from "../types/tauri";
+import type { InventoryUpdate } from "../types/inventory";
 import "./Overlay.css";
 
 interface ComponentRow {
